Create ClassOverview.txt
This document serves as a high-level overview of the integration details of each class and how they are used in the project.

This first draft is not 100% correct, but serves as a baseline because has 'mostly' correct info. Each class needs to be examined and it's overview refined and confirmed.